Move useParams to efi-ui/router
